fix: numeric-only locale dates, locale-aware chart axis, centered Consensus column
- Dates now use a numeric month everywhere (never a text month) while keeping the device's locale order — thumb-mark date, New Arrivals, and the masthead snapshot (he-IL → 18.7.2026, en-US → 7/18/2026, en-GB → 18/07/2026)
- Price-chart day axis was hardcoded M/D; now formats via the device locale (tz-safe, built from local Y/M/D parts)
- Consensus table column centered — header, the pill, and the ▲/●/▼ breakdown

+++ src/components/Masthead.tsx
@@ -10,7 +10,7 @@ interface MastheadProps {
 }
 
 const SNAPSHOT = new Date(meta.generatedAt).toLocaleString(DATE_LOCALE, {
-  month: "short", day: "numeric", year: "numeric",
+  month: "numeric", day: "numeric", year: "numeric",
   hour: "2-digit", minute: "2-digit", timeZone: "UTC", timeZoneName: "short",
 });
